+++ common/src/main/java/com/nemonotfound/nemos/copper/mixin/AbstractCowMixin.java
@@ -1,0 +1,39 @@
+package com.nemonotfound.nemos.copper.mixin;
+
+import com.nemonotfound.nemos.copper.item.Items;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.animal.AbstractCow;
+import net.minecraft.world.entity.animal.Animal;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.ItemUtils;
+import net.minecraft.world.level.Level;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+@Mixin(AbstractCow.class)
+public abstract class AbstractCowMixin extends Animal {
+
+    protected AbstractCowMixin(EntityType<? extends Animal> entityType, Level level) {
+        super(entityType, level);
+    }
+
+    @Inject(method = "mobInteract", at = @At("HEAD"), cancellable = true)
+    private void mobInteract(Player player, InteractionHand interactionHand, CallbackInfoReturnable<InteractionResult> cir) {
+        var itemStack = player.getItemInHand(interactionHand);
+
+        if (itemStack.is(Items.COPPER_BUCKET.get()) && !this.isBaby()) {
+            player.playSound(SoundEvents.COW_MILK, 1.0F, 1.0F);
+            ItemStack itemstack1 = ItemUtils.createFilledResult(itemStack, player, Items.COPPER_MILK_BUCKET.get().getDefaultInstance());
+            player.setItemInHand(interactionHand, itemstack1);
+
+            cir.setReturnValue(InteractionResult.SUCCESS);
+        }
+    }
+
+}