+++ common/src/main/java/com/nemonotfound/nemos/copper/mixin/TadpoleMixin.java
@@ -1,0 +1,16 @@
+package com.nemonotfound.nemos.copper.mixin;
+
+import com.nemonotfound.nemos.copper.interfaces.CopperBucketItemStackGetter;
+import com.nemonotfound.nemos.copper.item.Items;
+import net.minecraft.world.entity.animal.frog.Tadpole;
+import net.minecraft.world.item.ItemStack;
+import org.spongepowered.asm.mixin.Mixin;
+
+@Mixin(Tadpole.class)
+public class TadpoleMixin implements CopperBucketItemStackGetter {
+
+    @Override
+    public ItemStack nemosCopper$getCopperBucketItemStack() {
+        return new ItemStack(Items.COPPER_TADPOLE_BUCKET.get());
+    }
+}