+++ Changelog.md
@@ -1,7 +1,18 @@
 # Changelog v1.3
 
 ## Additions
-- Added copper bucket
+- Added copper buckets
+  - Copper Bucket
+  - Copper Water Bucket
+  - Copper Cod Bucket
+  - Copper Salmon Bucket
+  - Copper Tropical Fish Bucket
+  - Copper Pufferfish Bucket
+  - Copper Axolotl Bucket
+  - Copper Tadpole Bucket
+  - Copper Lava Bucket
+  - Copper Powder Snow Bucket
+  - Copper Milk Bucket
 
 ## Fixes
 - Copper minecarts drop normal minecart variants when broken